Move TypedDict declarations to separate file to avoid import issues
